requirements.txt
alabaster==0.7.12
amqp==2.5.2
attrs==19.1.0
Babel==2.7.0
billiard==3.6.1.0
celery==4.4.0
certifi==2019.3.9
chardet==3.0.4
codecov==2.0.15
commonmark==0.9.0
coverage==4.5.4
Django==2.2.5
django-nose==1.4.6
django-request-logging==0.7.0
djangorestframework==3.10.3
djangorestframework-bulk==0.2.1
docutils==0.15.2
env==0.1.0
future==0.17.1
idna==2.8
imagesize==1.1.0
importlib-metadata==1.3.0
Jinja2==2.10.1
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
mysqlclient==1.4.6
nose==1.3.7
packaging==19.1
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.7.5
python-dotenv==0.10.1
pytz==2018.9
recommonmark==0.6.0
requests==2.21.0
six==1.12.0
snowballstemmer==1.9.1
Sphinx==2.2.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
transitions==0.7.0
urllib3==1.24.1
vine==1.3.0
zipp==0.6.0
